inference_gui/requirements.txt
# Requirements for Swin Transformer Inference Scripts
# These should match the training environment

torch>=1.12.0
torchvision>=0.13.0
timm>=0.6.0
albumentations>=1.3.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.64.0
